Re-organize and document code
